Fix announcement bar inline script key not in sync with namespaced key
